fix: convert the velocity to m/s and do all subsequent computations with this value. Invert the sign of the recieved steering angle.
Convert the velocity to m/s.
Divide the psi prediction into smaller time steps. That way the
current steering angle would iteratively change the yaw angle.
Used actual latency time  (125-150ms) to account for the processing
time of the solver and the latency of the communication with
the simulator. Measured the latency using
std::chrono::system_clock::now().

Reviewer feedbacks